# tournois-sport-api/public/swagger.yaml
openapi: 3.0.3
info:
  title: Tournois Sport API
  description: API de gestion de tournois sportifs avec authentification, administration, joueurs, inscriptions, matchs et tournois.
  version: 1.0.0
servers:
  - url: http://127.0.0.1:8000
    description: Serveur local Symfony
...
paths:
...
  /register:
    post:
      tags: [Authentification]
      summary: Créer un utilisateur
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: Utilisateur créé
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: Données invalides
        '422':
          description: Email ou username déjà utilisé
  /api/login:
    post:
      tags: [Authentification]
      summary: Connexion utilisateur
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Connexion réussie
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Connexion réussie
                  user:
                    $ref: '#/components/schemas/User'
        '401':
          description: Identifiants invalides
  /api/me:
    get:
      tags: [Authentification]
      summary: Récupérer l'utilisateur connecté
      security:
        - sessionAuth: []
      responses:
        '200':
          description: Utilisateur connecté
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          description: Non authentifié
  /api/logout:
    post:
      tags: [Authentification]
      summary: Déconnexion utilisateur
      security:
        - sessionAuth: []
      responses:
        '200':
          description: Déconnexion réussie
        '401':
          description: Non authentifié
...
  /api/players/{id}:
    get:
      tags: [Joueurs]
      summary: Afficher un joueur
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '200':
          description: Joueur trouvé
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '404':
          description: Joueur introuvable
    put:
      tags: [Joueurs]
      summary: Modifier un joueur
      security:
        - sessionAuth: []
      parameters:
        - $ref: '#/components/parameters/Id'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserUpdateRequest'
      responses:
        '200':
          description: Joueur modifié
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          description: Non authentifié
        '403':
          description: Accès refusé
        '404':
          description: Joueur introuvable
    delete:
      tags: [Joueurs]
      summary: Supprimer un joueur
      security:
        - sessionAuth: []
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '204':
          description: Joueur supprimé
        '401':
          description: Non authentifié
        '403':
          description: Accès refusé
        '404':
          description: Joueur introuvable
...
  /admin/admin/dashboard:
    get:
      tags: [Administration API]
      summary: Dashboard admin API
      security:
        - sessionAuth: []
      responses:
        '200':
          description: Données du dashboard admin
        '401':
          description: Non authentifié
        '403':
          description: Accès réservé aux admins
  /admin/admin/registrations/{id}/confirm:
    put:
      tags: [Administration API]
      summary: Confirmer une inscription
      security:
        - sessionAuth: []
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '200':
          description: Inscription confirmée
        '401':
          description: Non authentifié
        '403':
          description: Accès réservé aux admins
        '404':
          description: Inscription introuvable
  /admin/admin/tournaments/{id}/winner:
    put:
      tags: [Administration API]
      summary: Définir le vainqueur d'un tournoi
      security:
        - sessionAuth: []
      parameters:
        - $ref: '#/components/parameters/Id'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [winnerId]
              properties:
                winnerId:
                  type: integer
                  example: 2
      responses:
        '200':
          description: Vainqueur défini
        '401':
          description: Non authentifié
        '403':
          description: Accès réservé aux admins
        '404':
          description: Tournoi ou utilisateur introuvable
  /admin/admin/sport-matchs/{id}/scores:
    put:
      tags: [Administration API]
      summary: Modifier les scores d'un match en tant qu'admin
      description: Lorsqu'un admin modifie les scores, aucune notification n'est envoyée.
      security:
        - sessionAuth: []
      parameters:
        - $ref: '#/components/parameters/Id'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SportMatchScoreUpdateRequest'
      responses:
        '200':
          description: Scores modifiés
        '401':
          description: Non authentifié
        '403':
          description: Accès réservé aux admins
        '404':
          description: Match introuvable
...
components:
  securitySchemes:
    sessionAuth:
      type: apiKey
      in: cookie
      name: PHPSESSID
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema:
        type: integer
      example: 1
...
  schemas:
    LoginRequest:
      type: object
      required:
        - emailAddress
        - password
      properties:
        emailAddress:
          type: string
          format: email
          example: chloe@example.com
        password:
          type: string
          example: password123
    RegisterRequest:
      type: object
      required:
        - lastName
        - firstName
        - username
        - emailAddress
        - password
      properties:
        lastName:
          type: string
          example: Dupont
        firstName:
          type: string
          example: Chloé
        username:
          type: string
          example: chloe
        emailAddress:
          type: string
          format: email
          example: chloe@example.com
        password:
          type: string
          example: password123
    UserUpdateRequest:
      type: object
      properties:
        lastName:
          type: string
          example: Dupont
        firstName:
          type: string
          example: Chloé
        username:
          type: string
          example: chloe
        emailAddress:
          type: string
          format: email
          example: chloe@example.com
        status:
          type: string
          example: actif
    User:
      type: object
      properties:
        id:
          type: integer
          example: 1
        lastName:
          type: string
          example: Dupont
        firstName:
          type: string
          example: Chloé
        username:
          type: string
          example: chloe
        emailAddress:
          type: string
          format: email
          example: chloe@example.com
        roles:
          type: array
          items:
            type: string
          example: [ROLE_USER]
        status:
          type: string
          example: actif
...
    SportMatchScoreUpdateRequest:
      type: object
      properties:
        scorePlayer1:
          type: integer
          nullable: true
          example: 10
        scorePlayer2:
          type: integer
          nullable: true
          example: 8
...